refactor: change architecture
