+++ main.js
@@ -3,6 +3,6 @@ import './src/sass/style.scss';
 import 'bootstrap';
 import 'bootstrap-icons/font/bootstrap-icons.css';
 
-import initPage from './src/ui/view/page';
+import initPage from './src/core/view/page';
 
 initPage();
