+++ src/core/game-figures/high/anchor.js
