+++ src/core/game-figures/high/death.js
