+++ src/core/game-figures/high/fish.js
