+++ src/core/game-figures/high/horse.js
