+++ src/core/game-figures/high/ram.js
