+++ src/core/game-figures/low/camel.js
