+++ src/core/game-figures/low/celtic-cross.js
