+++ src/core/game-figures/low/location-mark.js
