+++ src/core/game-figures/low/top-hat.js
