+++ src/core/game-figures/low/tower.js
