+++ src/core/game-figures/middle/bird.js
