+++ src/core/game-figures/middle/cat.js
