+++ src/core/game-figures/middle/deer.js
