+++ src/core/game-figures/middle/duck.js
