+++ src/core/game-figures/middle/house.js
