+++ src/core/game-figures/middle/snowman.js
