+++ src/core/game-figures/templates.js
@@ -1,19 +1,19 @@
-import locationMark from './game-figures/low/location-mark';
-import topHat from './game-figures/low/top-hat';
-import celticCross from './game-figures/low/celtic-cross';
-import tower from './game-figures/low/tower';
-import camel from './game-figures/low/camel';
-import bird from './game-figures/middle/bird';
-import cat from './game-figures/middle/cat';
-import deer from './game-figures/middle/deer';
-import duck from './game-figures/middle/duck';
-import house from './game-figures/middle/house';
-import snowman from './game-figures/middle/snowman';
-import death from './game-figures/high/death';
-import horse from './game-figures/high/horse';
-import fish from './game-figures/high/fish';
-import ram from './game-figures/high/ram';
-import anchor from './game-figures/high/anchor';
+import locationMark from './low/location-mark';
+import topHat from './low/top-hat';
+import celticCross from './low/celtic-cross';
+import tower from './low/tower';
+import camel from './low/camel';
+import bird from './middle/bird';
+import cat from './middle/cat';
+import deer from './middle/deer';
+import duck from './middle/duck';
+import house from './middle/house';
+import snowman from './middle/snowman';
+import death from './high/death';
+import horse from './high/horse';
+import fish from './high/fish';
+import ram from './high/ram';
+import anchor from './high/anchor';
 
 const templates = [
   {
