+++ src/core/shared/constants.js
@@ -1,0 +1,5 @@
+export const { BASE_URL } = import.meta.env;
+
+export const FILL_SOUND = `${BASE_URL}src/assets/sounds/fill.mp3`;
+export const CLEAN_SOUND = `${BASE_URL}src/assets/sounds/clean.wav`;
+export const X_SOUND = `${BASE_URL}src/assets/sounds/x.wav`;
