+++ src/core/shared/helpers.js
